refactor(runtime): move std modules to a separate sub-folder
