+++ dzx.ts
@@ -1,5 +1,3 @@
-/// <reference path="./types.d.ts" />
-
 import { dzx } from "./src/cli/mod.ts";
 
 if (import.meta.main) {
