+++ shell.ts
@@ -1,2 +1,1 @@
 export * from "./src/runtime/mod.ts";
-export * from "./src/std/mod.ts";
