+++ src/cli/lib/stream.ts
@@ -1,8 +1,8 @@
-import { ValidationError } from "../deps.ts";
+import { readAll, ValidationError } from "../deps.ts";
 import { bootstrapScript } from "./bootstrap.ts";
 
 export async function getModuleFromStdin(): Promise<string> {
-  const code = new TextDecoder().decode(await streams.readAll(Deno.stdin));
+  const code = new TextDecoder().decode(await readAll(Deno.stdin));
   if (!code) {
     throw new ValidationError(`Failed to read from stdin.`, { exitCode: 2 });
   }
