+++ src/runtime/deps.ts
@@ -1,50 +1,17 @@
-export * as async from "https://deno.land/std@0.140.0/async/mod.ts";
 export {
-  type DebouncedFunction,
   type Deferred,
   deferred,
-  type DelayOptions,
 } from "https://deno.land/std@0.140.0/async/mod.ts";
-export * as path from "https://deno.land/std@0.140.0/path/mod.ts";
-export type {
-  FormatInputPathObject,
-  GlobOptions,
-  GlobToRegExpOptions,
-  ParsedPath,
-} from "https://deno.land/std@0.140.0/path/mod.ts";
-export * as io from "https://deno.land/std@0.140.0/io/mod.ts";
-export { BufReader } from "https://deno.land/std@0.140.0/io/buffer.ts";
-export type {
-  ReadableStreamFromReaderOptions as IOReadableStreamFromReaderOptions,
-  ReadLineResult,
-  WritableStreamFromWriterOptions as IOWritableStreamFromWriterOptions,
-} from "https://deno.land/std@0.140.0/io/mod.ts";
-export * as streams from "https://deno.land/std@0.140.0/streams/mod.ts";
 export { writeAll } from "https://deno.land/std@0.140.0/streams/mod.ts";
-export type {
-  ReadableStreamFromReaderOptions,
-  WritableStreamFromWriterOptions,
-} from "https://deno.land/std@0.140.0/streams/mod.ts";
-export * as fs from "https://deno.land/std@0.140.0/fs/mod.ts";
-export type {
-  CopyOptions,
-  ExpandGlobOptions,
-  WalkEntry,
-  WalkOptions,
-} from "https://deno.land/std@0.140.0/fs/mod.ts";
-export * as log from "https://deno.land/std@0.140.0/log/mod.ts";
-export type {
-  FormatterFunction,
-  HandlerOptions,
-  LevelName,
-  LogConfig,
-  LogMode,
-} from "https://deno.land/std@0.140.0/log/mod.ts";
-export * as flags from "https://deno.land/std@0.140.0/flags/mod.ts";
-export type {
-  Args,
-  ParseOptions,
-} from "https://deno.land/std@0.140.0/flags/mod.ts";
+export { concat } from "https://deno.land/std@0.140.0/bytes/mod.ts";
+export {
+  bold,
+  brightBlue,
+  brightYellow,
+  red,
+  white,
+} from "https://deno.land/std@0.140.0/fmt/colors.ts";
+export { BufReader } from "https://deno.land/std@0.140.0/io/buffer.ts";
+export { join, sep } from "https://deno.land/std@0.140.0/path/mod.ts";
 export { colors } from "https://deno.land/x/cliffy@v0.24.2/ansi/colors.ts";
 export { default as shq } from "https://esm.sh/shq@1.0.2";
-export { concat } from "https://deno.land/std@0.140.0/bytes/mod.ts";
