--- src/runtime/globals.ts
@@ -1,5 +1,0 @@
-import * as globals from "./mod.ts";
-
-export function initGlobals() {
-  Object.assign(self, globals);
-}
