+++ src/std/deps.ts
@@ -1,0 +1,61 @@
+// std/async
+export * as async from "https://deno.land/std@0.140.0/async/mod.ts";
+export {
+  type DebouncedFunction,
+  type Deferred,
+  type DelayOptions,
+} from "https://deno.land/std@0.140.0/async/mod.ts";
+
+// std/path
+export * as path from "https://deno.land/std@0.140.0/path/mod.ts";
+export type {
+  FormatInputPathObject,
+  GlobOptions,
+  GlobToRegExpOptions,
+  ParsedPath,
+} from "https://deno.land/std@0.140.0/path/mod.ts";
+
+// std/io
+export * as io from "https://deno.land/std@0.140.0/io/mod.ts";
+export type {
+  ReadableStreamFromReaderOptions as IOReadableStreamFromReaderOptions,
+  ReadLineResult,
+  WritableStreamFromWriterOptions as IOWritableStreamFromWriterOptions,
+} from "https://deno.land/std@0.140.0/io/mod.ts";
+
+// std/streams
+export * as streams from "https://deno.land/std@0.140.0/streams/mod.ts";
+export { writeAll } from "https://deno.land/std@0.140.0/streams/mod.ts";
+export type {
+  ReadableStreamFromReaderOptions,
+  WritableStreamFromWriterOptions,
+} from "https://deno.land/std@0.140.0/streams/mod.ts";
+
+// std/fs
+export * as fs from "https://deno.land/std@0.140.0/fs/mod.ts";
+export type {
+  CopyOptions,
+  ExpandGlobOptions,
+  WalkEntry,
+  WalkOptions,
+} from "https://deno.land/std@0.140.0/fs/mod.ts";
+
+// std/log
+export * as log from "https://deno.land/std@0.140.0/log/mod.ts";
+export type {
+  FormatterFunction,
+  HandlerOptions,
+  LevelName,
+  LogConfig,
+  LogMode,
+} from "https://deno.land/std@0.140.0/log/mod.ts";
+
+// std/flags
+export * as flags from "https://deno.land/std@0.140.0/flags/mod.ts";
+export type {
+  Args,
+  ParseOptions,
+} from "https://deno.land/std@0.140.0/flags/mod.ts";
+
+// x/cliffy/ansi/colors: wrapper for std/flags
+export { colors } from "https://deno.land/x/cliffy@v0.24.2/ansi/colors.ts";
