+++ src/std/mod.ts
@@ -1,0 +1,1 @@
+export * from "./deps.ts";
