+++ test.ts
@@ -1,5 +1,3 @@
-/// <reference path="./types.d.ts" />
-
 import {
   assert,
   assertEquals,
